Rewrite logic about oauth2 grant
