+++ tests/tests_ridi_oauth2/test_grant.py
@@ -3,12 +3,12 @@ import unittest
 
 import requests_mock
 
-from ridi_oauth2.client.constants import OAuth2GrantType
 from ridi_oauth2.client.dtos import AuthorizationServerInfo, ClientInfo
-from ridi_oauth2.client.factory import OAuth2GrantFactory
+from ridi_oauth2.client.grant import Grant
+from ridi_oauth2.common.utils.string import generate_random_str
 
 
-class AuthorizationCodeGrantTestCase(unittest.TestCase):
+class GrantTestCase(unittest.TestCase):
     def setUp(self):
         self.client_info = ClientInfo(
             client_id='dummy_client_id', client_secret='dummy_client_secret', scope='all', redirect_uri='https://127.0.0.1:8000/callback'
@@ -16,20 +16,14 @@ class AuthorizationCodeGrantTestCase(unittest.TestCase):
         self.auth_server_info = AuthorizationServerInfo(
             authorization_url='https://127.0.0.1/oauth2/authorize', token_url='https://127.0.0.1/oauth2/token'
         )
-        self.grant_type = OAuth2GrantType.AUTHORIZATION_CODE
+        self.grant = Grant(client_info=self.client_info, auth_server_info=self.auth_server_info)
 
-    def test_request_authorize(self):
-        authorization_code_grant = OAuth2GrantFactory.get_grant(
-            grant_type=self.grant_type, client_info=self.client_info, auth_server_info=self.auth_server_info
-        )
-
-        authorize_url, _ = authorization_code_grant.get_authorization_url()
+    def test_authorize(self):
+        state = generate_random_str()
+        authorize_url = self.grant.authorize(state=state)
         self.assertIn(self.auth_server_info.authorization_url, authorize_url)
 
-    def test_get_token(self):
-        authorization_code_grant = OAuth2GrantFactory.get_grant(
-            grant_type=self.grant_type, client_info=self.client_info, auth_server_info=self.auth_server_info
-        )
+    def test_code(self):
         dummy_code = 'this.is.dummy.code'
 
         with requests_mock.Mocker() as m:
@@ -42,32 +36,13 @@ class AuthorizationCodeGrantTestCase(unittest.TestCase):
                 'scope': 'all',
             }))
 
-            result = authorization_code_grant.get_access_token(code=dummy_code)
+            result = self.grant.code(code=dummy_code)
 
             self.assertIsNotNone(result.access_token.token)
             self.assertIsNotNone(result.refresh_token.token)
             self.assertEqual(result.scope, self.client_info.scope)
 
-
-class RefreshTokenGrantTestCase(unittest.TestCase):
-    def setUp(self):
-        self.client_info = ClientInfo(client_id='dummy_client_id', client_secret='dummy_client_secret', scope='all', )
-        self.auth_server_info = AuthorizationServerInfo(token_url='https://127.0.0.1/oauth2/token')
-        self.grant_type = OAuth2GrantType.REFRESH_TOKEN
-
-    def test_not_implement_authorize(self):
-        refresh_token_grant = OAuth2GrantFactory.get_grant(
-            grant_type=self.grant_type, client_info=self.client_info, auth_server_info=self.auth_server_info
-        )
-
-        with self.assertRaises(NotImplementedError):
-            refresh_token_grant.get_authorization_url()
-
-    def test_get_token(self):
-        refresh_token_grant = OAuth2GrantFactory.get_grant(
-            grant_type=self.grant_type, client_info=self.client_info, auth_server_info=self.auth_server_info
-        )
-
+    def test_refresh(self):
         dummy_refresh_token = 'this.is.refresh.token.'
 
         with requests_mock.Mocker() as m:
@@ -80,7 +55,7 @@ class RefreshTokenGrantTestCase(unittest.TestCase):
                 'scope': 'all',
             }))
 
-            result = refresh_token_grant.get_access_token(refresh_token=dummy_refresh_token)
+            result = self.grant.refresh(refresh_token=dummy_refresh_token)
 
             self.assertIsNotNone(result.access_token)
             self.assertIsNotNone(result.refresh_token)
